(fix) not deleting items if first slot is lower
Found this bug in my server, where, in case you had 1 slot with an amount under the total amount to remove, (for example, 1 in slot1 and 4 in slot 4, try to remove 5), it would return true but not remove anything. This fixes it by removing the first slot and continuing without updating.